Vote with a unique code
Everyone gets sent a unique voting code which allows them to vote once. After using it, it expires. This way, an extensive and uncertain scan for proxies is no longer necessary.
The generate-codes script generates random vote codes; mysql_import puts them in a MySQL database. When someone votes, the code is removed from the database. Votes remain anonymous.
In theory, one could brute-force the vote codes. However, it is very unlikely that they would ever find one because of 1. the use of Google reCAPTCHA, and 2. the complexity of the codes.