Made some styling changes to poll.php for login and voting page
Included some basic markup for the login page. The login page now contains a centered login screen with a nice text field and button for the user to enter their code.

Included some basic markup for the voting page, all the possible votes are now displayed in boxes with a description of the item below it. Everything is alligned in the middle and looks a bit better this way. I moved the options to a different file where you can modify an array of items with a name and a description key. This file is then included and the options can be read out of the array. Since the options started with id=1 I left it that way.