Clean up options for-loop
Use the fancy colon for-loop instead of intertwining a for-loop and an enhanced for-loop. The array in options.php is a multidimensional array, so we can do things like $options[$i]['name'].

The vote options now start with 0. The order in which the options are in options.php DOES matter now. The database will also get messed up if the order changes after it has been used. This could be solved by putting a vote ID in options.php, but I think that might be too complicated for the people who put the options in there.