Put css in php/; use relative links instead of absolute
The stylesheets are in php/ now: php/ is actually the webserver's root. Maybe it should be named web/ or something. This is because the shell scripts are intended to be run when an admin is setting up the poll; the webserver is not interfering with it.

Also, I replaced the absolute links to the stylesheets (like /foundation/...) to relative ones, so they can be found regardless of where the program is placed in the webserver.