Login with plain password in POST request
Instead of trying to login via /ProcessLogin.aspx with the password as a hash in a GET request, we can just use the normal login to avoid having to imitate the server's spooky hashing method for the password.
So basically, we're just logging in like any normal user would do, except that the login is our first communication with the server where we also get the SessionId cookie. After our SessionId has been authenticated by logging in, we should be able to use it to send internal messages.